nanogui type casting warnings in screen fixed
bugs with returning an existing map element fixed
- hash was not being cut down to the bit depth of HshType::hsh
- a new KV was being created instead of returning the existing KV

+++ nuklear/nanogui/src/screen.cpp
@@ -413,29 +413,29 @@ void Screen::drawWidgets() {
             int h = (int)( (bounds[2] - bounds[0]) / 2 );
             if (h > tooltipWidth / 2) {
                 nvgTextAlign(mNVGContext, NVG_ALIGN_CENTER | NVG_ALIGN_TOP);
-                nvgTextBoxBounds(mNVGContext, pos.x(), pos.y(), tooltipWidth,
+                nvgTextBoxBounds(mNVGContext, (float)pos.x(), (float)pos.y(), (float)tooltipWidth,
                                 widget->tooltip().c_str(), nullptr, bounds);
 
-                h = (bounds[2] - bounds[0]) / 2;
+                h = (int)( (bounds[2] - bounds[0]) / 2 );
             }
             nvgGlobalAlpha(mNVGContext,
                            std::min(1.0, 2 * (elapsed - 0.5f)) * 0.8);
 
             nvgBeginPath(mNVGContext);
             nvgFillColor(mNVGContext, Color(0, 255));
-            nvgRoundedRect(mNVGContext, bounds[0] - 4 - h, bounds[1] - 4,
-                           (int) (bounds[2] - bounds[0]) + 8,
-                           (int) (bounds[3] - bounds[1]) + 8, 3);
+            nvgRoundedRect(mNVGContext, bounds[0] - 4 - h ,  bounds[1] - 4,
+                           bounds[2] - bounds[0] + 8,
+                           bounds[3] - bounds[1] + 8, 3);
 
             int px = (int) ((bounds[2] + bounds[0]) / 2) - h;
-            nvgMoveTo(mNVGContext, px, bounds[1] - 10);
-            nvgLineTo(mNVGContext, px + 7, bounds[1] + 1);
-            nvgLineTo(mNVGContext, px - 7, bounds[1] + 1);
+            nvgMoveTo(mNVGContext, (float)px, bounds[1] - 10);
+            nvgLineTo(mNVGContext, px + 7.f, bounds[1] + 1);
+            nvgLineTo(mNVGContext, px - 7.f, bounds[1] + 1);
             nvgFill(mNVGContext);
 
             nvgFillColor(mNVGContext, Color(255, 255));
             nvgFontBlur(mNVGContext, 0.0f);
-            nvgTextBox(mNVGContext, pos.x() - h, pos.y(), tooltipWidth,
+            nvgTextBox(mNVGContext, (float)(pos.x() - h), (float)pos.y(), (float)tooltipWidth,
                        widget->tooltip().c_str(), nullptr);
         }
     }
@@ -474,7 +474,7 @@ bool Screen::cursorPosCallbackEvent(double x, double y) {
     Vector2i p((int) x, (int) y);
 
 #if defined(_WIN32) || defined(__linux__)
-    p /= mPixelRatio;
+    p /= (int)mPixelRatio;
 #endif
 
     bool ret = false;
@@ -609,7 +609,7 @@ bool Screen::resizeCallbackEvent(int, int) {
     glfwGetWindowSize(mGLFWWindow, &size[0], &size[1]);
 
 #if defined(_WIN32) || defined(__linux__)
-    size /= mPixelRatio;
+    size /= (int)mPixelRatio;
 #endif
 
     if (mFBSize == Vector2i(0, 0) || size == Vector2i(0, 0))
